+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/Youtube.java
@@ -1,10 +1,10 @@
-package de.slimecloud.slimeball.features.youtube;
+package de.slimecloud.slimeball.features.alerts.youtube;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import de.slimecloud.slimeball.features.youtube.model.SearchResult;
-import de.slimecloud.slimeball.features.youtube.model.Video;
+import de.slimecloud.slimeball.features.alerts.youtube.model.SearchResult;
+import de.slimecloud.slimeball.features.alerts.youtube.model.Video;
 import de.slimecloud.slimeball.main.Main;
 import de.slimecloud.slimeball.main.SlimeBot;
 import lombok.RequiredArgsConstructor;
