+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/YoutubeConfig.java
@@ -1,4 +1,4 @@
-package de.slimecloud.slimeball.features.youtube;
+package de.slimecloud.slimeball.features.alerts.youtube;
 
 import de.slimecloud.slimeball.config.engine.Required;
 import lombok.Getter;
