+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/model/ChannelSnippet.java
@@ -1,0 +1,5 @@
+package de.slimecloud.slimeball.features.alerts.youtube.model;
+
+public record ChannelSnippet(String title) {
+
+}
