+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/model/ResourceId.java
@@ -1,0 +1,4 @@
+package de.slimecloud.slimeball.features.alerts.youtube.model;
+
+public record ResourceId(String videoId) {
+}
