+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/model/SearchResultSnippet.java
@@ -1,4 +1,4 @@
-package de.slimecloud.slimeball.features.youtube.model;
+package de.slimecloud.slimeball.features.alerts.youtube.model;
 
 public record SearchResultSnippet(String channelId, String channelTitle, String title, String description, String liveBroadcastContent) {
 }
