+++ src/main/java/de/slimecloud/slimeball/features/alerts/youtube/model/VideoSnippet.java
@@ -1,3 +1,3 @@
-package de.slimecloud.slimeball.features.youtube.model;
+package de.slimecloud.slimeball.features.alerts.youtube.model;
 
 public record VideoSnippet(String description, String title, String channelId, String channelTitle, String liveBroadcastContent) {}
