--- src/main/java/de/slimecloud/slimeball/features/youtube/model/ChannelSnippet.java
@@ -1,5 +1,0 @@
-package de.slimecloud.slimeball.features.youtube.model;
-
-public record ChannelSnippet(String title) {
-
-}
