--- src/main/java/de/slimecloud/slimeball/features/youtube/model/ResourceId.java
@@ -1,4 +1,0 @@
-package de.slimecloud.slimeball.features.youtube.model;
-
-public record ResourceId(String videoId) {
-}
